completed the creatives.js file
