+++ models/creatives.js
@@ -1,36 +1,30 @@
-const sequelize = require("./index.js").sequelize;
-const test_db = require("../config/config.json");
-const mysql = require("mysql2");
+// Creating our User model 
 
-const Creatives = sequelize.define(
-  "creatives",
-  {
+module.exports = function(sequelize, DataTypes) {
+  var Creatives = sequelize.define("Creative", {
+    // The email cannot be null, and must be a proper email before creation
     id: {
-      type: sequelize.INTEGER,
-      primaryKey: true,
-      autoIncrement: true,
-    },
-    name: {
-      type: sequelize.STRING,
-      allowNull: false,
-    },
-    //   type: { type: sequelize.STRING }, move to media table
-    bio: {
-      type: sequelize.STRING,
-      allowNull: false,
-    },
-    post_id: {
-      type: sequelize.INTEGER,
-      allowNull: false,
-    },
-    media_id: {
-      type: sequelize.INTEGER,
-      allowNull: false,
-    },
-  }
-  // { sequelize }
-);
-sequelize.models.creatives;
-
-return Creatives;
-module.exports = Creatives;
+        type: sequelize.INTEGER,
+        primaryKey: true,
+        autoIncrement: true,
+      },
+      name: {
+        type: sequelize.STRING,
+        allowNull: false,
+      },
+      //   type: { type: sequelize.STRING }, move to media table
+      bio: {
+        type: sequelize.STRING,
+        allowNull: false,
+      },
+      post_id: {
+        type: sequelize.INTEGER,
+        allowNull: false,
+      },
+      media_id: {
+        type: sequelize.INTEGER,
+        allowNull: false,
+      },
+  });
+  return Creatives;
+};
