+++ package.json
@@ -20,9 +20,13 @@
   },
   "homepage": "https://github.com/pakmk/gt-project2#readme",
   "dependencies": {
+    "bcrypt": "^5.0.0",
     "express": "^4.17.1",
     "express-handlebars": "^3.1.0",
+    "express-session": "^1.17.1",
     "mysql2": "^2.1.0",
+    "passport": "^0.4.1",
+    "passport-local": "^1.0.0",
     "sequelize": "^6.3.3"
   }
 }
