+++ server.js
@@ -1,56 +1,30 @@
-/**
- * REQUIRE NPM PACKAGES
- * REQUIRE EXTERNAL FILES
- */
-const express = require("express");
-const exphbs = require("express-handlebars");
+// Requiring necessary npm packages
+var express = require("express");
+var session = require("express-session");
+// Requiring passport as we've configured it
+var passport = require("./config/passport");
 
-/**
- * DEFINE VARIABLES
- */
-const PORT = process.env.PORT || 8080;
-const app = express();
-const db = require("./models/creatives.js");
-const ViewsController = require("./controllers/viewsController.js");
-const APIController = require("./controllers/apiController");
-const UsersController = require("./controllers/usersController");
+// Setting up port and requiring models for syncing
+var PORT = process.env.PORT || 8080;
+var db = require("./models");
 
-/**
- * MIDDLEWARE
- */
-// Parse application body as JSON
+// Creating express app and configuring middleware needed for authentication
+var app = express();
 app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
-// Handlebars setup
-app.engine("handlebars", exphbs({ defaultLayout: "main" }));
-app.set("view engine", "handlebars");
-
-/**
- * VIEW ROUTES
- * API ROUTES
- */
-// Serve static content for the app from the "public" directory in the application directory.
 app.use(express.static("public"));
+// We need to use sessions to keep track of our user's login status
+app.use(session({ secret: "keyboard cat", resave: true, saveUninitialized: true }));
+app.use(passport.initialize());
+app.use(passport.session());
 
-// Routes
-app.use(ViewsController);
-app.use(APIController);
-app.use("/api/users", UsersController);
+// Requiring our routes
+require("./routes/html-routes.js")(app);
+require("./routes/api-routes.js")(app);
 
-/**
- * DB Connection
- * APP LISTEN
- */
-db.sequelize
-  .sync()
-  .then(() => {
-    // Start our server so that it can begin listening to client requests.
-    app.listen(PORT, function () {
-      // Log (server-side) when our server has started
-      console.log(`Server listening on: http://localhost:${PORT}`);
-    });
-  })
-  .catch((err) => {
-    console.log("Error connecting to the database.");
-    console.log(err);
+// Syncing our database and logging a message to the user upon success
+db.sequelize.sync().then(function() {
+  app.listen(PORT, function() {
+    console.log("==> 🌎  Listening on port %s. Visit http://localhost:%s/ in your browser.", PORT, PORT);
   });
+});
